requirements.txt
editdistance==0.4
Keras==2.2.0
tensorflow==1.5.0
matplotlib
pillow
scikit-learn
h5py
